refactor: revert changes to appIcon.tsx and imports
